fixed seed issue
